requirements.txt
academic==0.7.0
asgiref==3.3.1
beautifulsoup4==4.9.3
bibtexparser==1.1.0
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-bootstrap-v5==1.0.0
django-ckeditor==6.0.0
django-colorfield==0.4.1
django-crispy-forms==1.11.1
django-jalali==4.1.2
django-js-asset==1.2.2
django-quill-editor==0.1.21
django-summernote==0.8.11.6
django-tinymce==3.2.0
djangorestframework==3.12.2
future==0.18.2
idna==2.10
importlib-metadata==2.1.1
jdatetime==3.6.2
Pillow==8.1.0
pyparsing==2.4.7
pytz==2021.1
requests==2.25.1
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.2
soupsieve==2.2
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.3
zipp==3.4.0
